+++ src/main/webapp/resources/css/schedule/scheduleCalendar.css
@@ -9,8 +9,6 @@
 }
 
 div.scheduleCalendar {
-	width: 80%;
-	max-width: 1280px;
 	margin: auto;
 }
