+++ src/main/webapp/resources/js/schedule/scheduleCalendar.js
@@ -31,6 +31,8 @@ $(document).ready(() => {
 		]
 	});
 	calendar.render();
-
+	
+	
+	console.log(test);
 
 });
